PhyloFun is now running with the new Rcpp generation of the bayesian conditional probability tables.
PhyloFun is now a "true" R package and can be installed with R CMD INSTALL PhyloFun/
Moved some files around and added Description and Namespace files.
Also debugged some functions that accessed members of XStringSet, somehow on debian the functions worked but on Mac OS X not. Had to use the double square bracket accessor and length on S4 class AAString instead of nchar and converting it to character vector of length 1…

Still missing:
 * remove project.file.path calls in all scripts.
 * adjust exec scripts to require PhyloFun and if needed load PhyloFun's data
 * Move all tests into test functions
 * Generate Rdoc from the functions comments. Maybe by using Vim macros.